Move queries to own class
